Part tracks managed resources, not prefab
Use of PartUpgrades to control the available types list means the prefab will not have the correct view of the managed resources.  That view is dynamic based on the final typesAvailable list, which won't be known until the Editor ot Flight scenes.
Move the tracking out of a static HashSet in MFSSettings.